Abstracted method monitor_transaction()
