+++ anydex/wallet/bitcoinlib/bitcoinlib_wallets.py
@@ -1,7 +1,6 @@
 import os
 import time
 
-from asyncio import Future
 from binascii import hexlify
 from configparser import ConfigParser
 
@@ -212,25 +211,6 @@ class BitcoinlibWallet(Wallet):
 
     def precision(self):
         return 8       # The precision for LTC, BTC and DASH is the same.
-
-    def monitor_transaction(self, txid):
-        """
-        Monitor a given transaction ID. Returns a Deferred that fires when the transaction is present.
-        """
-        monitor_future = Future()
-
-        async def monitor():
-            transactions = await self.get_transactions()
-            for transaction in transactions:
-                if transaction['id'] == txid:
-                    self._logger.debug("Found transaction with id %s", txid)
-                    monitor_future.set_result(None)
-                    monitor_task.cancel()
-
-        self._logger.debug("Start polling for transaction %s", txid)
-        monitor_task = self.register_task(f"{self.network}_poll_{txid}", monitor, interval=5)
-
-        return monitor_future
 
     def is_testnet(self):
         return self.testnet
